# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SoftKeys Store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  icon: string;
  productCount?: number;
}

export type ProductType = (typeof ProductType)[keyof typeof ProductType];

export const ProductType = {
  key: "key",
  subscription: "subscription",
} as const;

export interface Product {
  id: number;
  name: string;
  slug: string;
  description: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  categoryId: number;
  categoryName: string;
  type: ProductType;
  /**
   * e.g. '1 year', '3 years', 'lifetime'
   * @nullable
   */
  subscriptionDuration?: string | null;
  /** e.g. 'Windows', 'Mac', 'Cross-platform' */
  platform: string;
  inStock: boolean;
  featured?: boolean;
  onSale?: boolean;
  rating: number;
  reviewCount: number;
  imageUrl: string;
  /**
   * e.g. 'Best Seller', 'Hot Deal', 'New'
   * @nullable
   */
  badge?: string | null;
}

export interface StoreStats {
  totalProducts: number;
  totalCategories: number;
  totalKeysSold: number;
  satisfiedCustomers: number;
}

export interface OrderItem {
  productId: number;
  quantity: number;
}

export type OrderInputPaymentMethod =
  (typeof OrderInputPaymentMethod)[keyof typeof OrderInputPaymentMethod];

export const OrderInputPaymentMethod = {
  card: "card",
  paypal: "paypal",
  crypto: "crypto",
} as const;

export interface OrderInput {
  items: OrderItem[];
  customerEmail: string;
  customerName: string;
  paymentMethod: OrderInputPaymentMethod;
  /** @nullable */
  couponCode?: string | null;
}

export interface OrderItemFull {
  productId: number;
  productName: string;
  quantity: number;
  unitPrice: number;
  /** @nullable */
  licenseKey?: string | null;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  completed: "completed",
  failed: "failed",
} as const;

export interface Order {
  id: number;
  customerEmail: string;
  customerName: string;
  items: OrderItemFull[];
  total: number;
  /** @nullable */
  discount?: number | null;
  status: OrderStatus;
  paymentMethod: string;
  createdAt: string;
}

export type ListProductsParams = {
  /**
   * @nullable
   */
  categoryId?: number | null;
  /**
   * @nullable
   */
  search?: string | null;
  /**
   * @nullable
   */
  featured?: boolean | null;
  /**
   * @nullable
   */
  onSale?: boolean | null;
  /**
   * @nullable
   */
  type?: ListProductsType;
};

export type ListProductsType =
  | (typeof ListProductsType)[keyof typeof ListProductsType]
  | null;

export const ListProductsType = {
  key: "key",
  subscription: "subscription",
} as const;