# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SoftKeys Store API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItemFull } from "./orderItemFull";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  customerEmail: string;
  customerName: string;
  items: OrderItemFull[];
  total: number;
  /** @nullable */
  discount?: number | null;
  status: OrderStatus;
  paymentMethod: string;
  createdAt: string;
}