# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SoftKeys Store API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputPaymentMethod } from "./orderInputPaymentMethod";
import type { OrderItem } from "./orderItem";

export interface OrderInput {
  items: OrderItem[];
  customerEmail: string;
  customerName: string;
  paymentMethod: OrderInputPaymentMethod;
  /** @nullable */
  couponCode?: string | null;
}