# lib/api-zod/src/generated/types/orderItemFull.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SoftKeys Store API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItemFull {
  productId: number;
  productName: string;
  quantity: number;
  unitPrice: number;
  /** @nullable */
  licenseKey?: string | null;
}