# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SoftKeys Store API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductType } from "./productType";

export interface Product {
  id: number;
  name: string;
  slug: string;
  description: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  categoryId: number;
  categoryName: string;
  type: ProductType;
  /**
   * e.g. '1 year', '3 years', 'lifetime'
   * @nullable
   */
  subscriptionDuration?: string | null;
  /** e.g. 'Windows', 'Mac', 'Cross-platform' */
  platform: string;
  inStock: boolean;
  featured?: boolean;
  onSale?: boolean;
  rating: number;
  reviewCount: number;
  imageUrl: string;
  /**
   * e.g. 'Best Seller', 'Hot Deal', 'New'
   * @nullable
   */
  badge?: string | null;
}