# lib/api-zod/src/generated/types/storeStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SoftKeys Store API
 * OpenAPI spec version: 0.1.0
 */

export interface StoreStats {
  totalProducts: number;
  totalCategories: number;
  totalKeysSold: number;
  satisfiedCustomers: number;
}